@@ -2,15 +2,14 @@ package org.springframework.data.elasticsearch.client.elc;
 
 import co.elastic.clients.elasticsearch.ElasticsearchClient;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.data.elasticsearch.core.ExtendedSearchOperations;
 import org.springframework.data.elasticsearch.core.convert.ElasticsearchConverter;
 
-@Configuration
 public interface ExtendedSearchConfiguration {
 
     /**
-     * Creates {@link ExtendedSearchOperations}.
+     * Creates a {@link ExtendedSearchOperations} implementation using an
+     * {@link co.elastic.clients.elasticsearch.ElasticsearchClient}.
      *
      * @return never {@literal null}.
      */
